+++ iosApp/iosApp/SceneDelegate.swift
@@ -13,10 +13,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
         // Create the SwiftUI view that provides the window contents.
 
-        let contentView = NavigationView {
-            ContentView()
-        }
-        .navigationViewStyle(StackNavigationViewStyle())
+        let contentView = NavigationView { HomeView() }
+            .navigationViewStyle(StackNavigationViewStyle())
 
         // Use a UIHostingController as window root view controller.
         if let windowScene = scene as? UIWindowScene {
