+++ iosApp/iosApp/Views/ActivityIndicator.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+struct ActivityIndicator: UIViewRepresentable {
+
+    func makeUIView(context: Context) -> UIActivityIndicatorView {
+        UIActivityIndicatorView()
+    }
+    
+    func updateUIView(_ uiView: UIActivityIndicatorView, context: Context) {
+        uiView.startAnimating()
+    }
+}
